Random Leaf (Vertically Biased Selection) Mode Added to Point Cloud Baker Utility
This addition allows not only grass created on terrain, as in the
GrassWind Sample but also VFX leaves to be generating on tree trunks.

Added a mode to the VFX point cache baker utility to select points that
are only in the top half of the mesh. Additionally points above the
half way mark are added at a greater probability towards the top of the
mesh.

Used for generating points to attach VFX leaves onto.